refactor: OutputView 파일에서 BlockProcessor 함수 분리, validators 파일 이름 변경
